Move createUser to data
